Add nfs mode
To mount rootfs via 9p, we must use two volumes (one for rootfs and one
for hostfs). To mount them we need to use intermediate init to mount
hostfs and be able to use utilities from the host system.

The other problem is that 9p needs to be specifically enabled in the
kernel and there is no default support in the kernel.

To solve this, it is possible to use a different rootfs mount method. We
can use nfsroot to mount it. This method has been around for a long time
and is very stable.

To combine rootfs and hostfs we implement our own nfs server.

Signed-off-by: Alexey Gladkov <legion@kernel.org>